@@ -1,13 +1,14 @@
-const path = require("path")
+import { relative } from "path"
+import process from "process"
 
 const buildEslintCommand = (filenames) =>
   `next lint --fix --file ${filenames
-    .map((f) => path.relative(process.cwd(), f))
+    .map((f) => relative(process.cwd(), f))
     .join(" --file ")}`
 
 const prettierCommand = "prettier --write"
 
-module.exports = {
+export default {
   "*.{js,jsx,ts,tsx}": [prettierCommand, buildEslintCommand],
   "*.{json,css,md}": [prettierCommand],
 }
