@@ -1,5 +1,5 @@
 import type { Metadata } from "next"
-import localFont from "next/font/local"
+import { Geist, Geist_Mono } from "next/font/google"
 
 import { cn } from "@/lib/utils"
 import { Providers } from "@/components/providers"
@@ -7,15 +7,14 @@ import { SiteHeader } from "@/components/site-header"
 
 import "./globals.css"
 
-const geistSans = localFont({
-  src: "./fonts/GeistVF.woff",
+const geistSans = Geist({
   variable: "--font-geist-sans",
-  weight: "100 900",
+  subsets: ["latin"],
 })
-const geistMono = localFont({
-  src: "./fonts/GeistMonoVF.woff",
+
+const geistMono = Geist_Mono({
   variable: "--font-geist-mono",
-  weight: "100 900",
+  subsets: ["latin"],
 })
 
 export const metadata: Metadata = {
@@ -29,16 +28,19 @@ export default function RootLayout({
   children: React.ReactNode
 }>) {
   return (
-    <html lang="en" suppressHydrationWarning>
+    <html lang="en" suppressHydrationWarning className="scroll-pt-14">
       <body
         className={cn(
-          "min-h-screen antialiased",
           geistSans.variable,
-          geistMono.variable
+          geistMono.variable,
+          "min-h-screen antialiased"
         )}
       >
         <Providers>
-          <div className="relative flex min-h-dvh flex-col bg-white bg-[radial-gradient(100%_50%_at_50%_0%,rgba(0,163,255,0.13)_0,rgba(0,163,255,0)_50%,rgba(0,163,255,0)_100%)] dark:bg-neutral-950 dark:bg-[radial-gradient(ellipse_80%_80%_at_50%_-20%,rgba(120,119,198,0.3),rgba(255,255,255,0))]">
+          <div className="fixed left-0 top-0 -z-10 size-full">
+            <div className="absolute top-0 z-[-2] h-screen w-screen bg-white bg-[radial-gradient(100%_50%_at_50%_0%,rgba(0,163,255,0.13)_0,rgba(0,163,255,0)_50%,rgba(0,163,255,0)_100%)] dark:bg-neutral-950 dark:bg-[radial-gradient(ellipse_80%_80%_at_50%_-20%,rgba(120,119,198,0.3),rgba(255,255,255,0))]" />
+          </div>
+          <div className="relative flex min-h-dvh flex-col">
             <SiteHeader />
             <main className="flex-1">{children}</main>
           </div>
