@@ -30,7 +30,7 @@ export function MainNav() {
         href="/about"
         className={cn(
           "hidden text-sm font-medium transition-colors hover:text-primary sm:inline-block",
-          pathname === "/blog" ? "text-foreground" : "text-foreground/60"
+          pathname === "/about" ? "text-foreground" : "text-foreground/60"
         )}
       >
         About
